Create cross-platform Conda environment file
This environment file is very convenient for Conda users to create an Pommerman environment by `conda env create -f env.yml`.

And Conda is much better because it can manage versions of Python and installation of CUDA toolkits.

This file is tested on Windows and Linux. And versions inside is consistent with those in `requirements.txt`.